+++ python/sglang/srt/models/deepseek_common/attention_forward_methods/__init__.py
@@ -1,7 +1,13 @@
 from .forward_methods import AttnForwardMethod
 from .forward_mha import DeepseekMHAForwardMixin
+from .forward_mla import DeepseekMLAForwardMixin
+from .forward_mla_fused_rope_cpu import DeepseekMLACpuForwardMixin
+from .forward_mla_fused_rope_rocm import DeepseekMLARocmForwardMixin
 
 __all__ = [
     "AttnForwardMethod",
     "DeepseekMHAForwardMixin",
+    "DeepseekMLACpuForwardMixin",
+    "DeepseekMLAForwardMixin",
+    "DeepseekMLARocmForwardMixin",
 ]
